pyproject.toml
[tool.poetry]
package-mode = false
name = "ai-friend-chatbot"
version = "0.1.0"
description = "AI friend chatbot built with AutoGen and Mem0"
authors = ["Lennex Zinyando <brizdigital@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
groq = "^0.11.0"
chromadb = "^0.5.5"
gradio = "^4.43.0"
pyautogen = "^0.3.0"
mem0ai = "^0.1.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
